requirements.txt
Django==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
cx_Oracle==8.3.0
Pillow==10.0.0
requests==2.31.0
PyJWT==2.8.0
uritemplate==4.1.1
inflection==0.5.1